+++ sdk/ai/Azure.AI.Extensions.OpenAI/src/Telemetry/OpenTelemetryResponseScope.cs
@@ -5,6 +5,7 @@ using System;
 using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Metrics;
 using System.IO;
 using System.Text;
@@ -158,6 +159,7 @@ namespace Azure.AI.Extensions.OpenAI.Telemetry
             return scope;
         }
 
+        [Experimental("SCME0001")]
         internal static OpenTelemetryResponseScope Start(CreateResponseOptions options, Uri endpoint, string defaultModelName)
         {
             if (!IsEnabled)
@@ -172,6 +174,7 @@ namespace Azure.AI.Extensions.OpenAI.Telemetry
             return scope;
         }
 
+        [Experimental("SCME0001")]
         internal static void ExtractOptionsContext(
             CreateResponseOptions options,
             string defaultModelName,